part1 semi comp

+++ ft_memchr.c
@@ -6,18 +6,20 @@
 /*   By: hohnuki <hohnuki@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/06 19:44:54 by hohnuki           #+#    #+#             */
-/*   Updated: 2021/10/06 20:29:34 by hohnuki          ###   ########.fr       */
+/*   Updated: 2021/10/07 19:07:16 by hohnuki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memchr(const void *s, int c, int n)
+#include "libft.h"
+
+void	*ft_memchr(const void *s, int c, size_t n)
 {
 	int				i;
 	unsigned char	*ptr;
 
 	i = 0;
 	ptr = (unsigned char *)s;
-	while (i < n)
+	while (i < (int)n)
 	{
 		if (ptr[i] == (char)c)
 			return (&ptr[i]);
