+++ ft_memcmp.c
@@ -6,24 +6,26 @@
 /*   By: hohnuki <hohnuki@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/06 20:20:32 by hohnuki           #+#    #+#             */
-/*   Updated: 2021/10/06 20:29:03 by hohnuki          ###   ########.fr       */
+/*   Updated: 2021/10/07 19:11:37 by hohnuki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_memcmp(const void *s1, const void *s2, unsigned int size)
+#include "libft.h"
+
+int	ft_memcmp(const void *s1, const void *s2, size_t size)
 {
-	unsigned int	i;
+	int	i;
 	unsigned char	*ptr1;
 	unsigned char	*ptr2;
 
 	i = 0;
 	ptr1 = (unsigned char *)s1;
 	ptr2 = (unsigned char *)s2;
-	while (ptr1[i] == ptr2[i] && ptr1[i] != '\0' && i < size)
+	while (ptr1[i] == ptr2[i] && ptr1[i] != '\0' && i < (int)size)
 	{
 		i++;
 	}
-	if (i == size)
+	if (i == (int)size)
 		return (0);
 	else
 		return (ptr1[i] - ptr2[i]);
