+++ ft_memcpy.c
@@ -6,11 +6,13 @@
 /*   By: hohnuki <hohnuki@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/06 15:43:10 by hohnuki           #+#    #+#             */
-/*   Updated: 2021/10/06 16:37:32 by hohnuki          ###   ########.fr       */
+/*   Updated: 2021/10/07 18:31:16 by hohnuki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memcpy(void *buf1, const void *buf2, int n)
+#include "libft.h"
+
+void	*ft_memcpy(void *buf1, const void *buf2, size_t n)
 {
 	int				i;
 	unsigned char	*ptr1;
@@ -19,7 +21,7 @@ void	*ft_memcpy(void *buf1, const void *buf2, int n)
 	i = 0;
 	ptr1 = (unsigned char *)buf1;
 	ptr2 = (unsigned char *)buf2;
-	while (i < n)
+	while (i < (int)n)
 	{
 		ptr1[i] = ptr2[i];
 		i++;
