+++ ft_memmove.c
@@ -6,11 +6,13 @@
 /*   By: hohnuki <hohnuki@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/06 15:59:47 by hohnuki           #+#    #+#             */
-/*   Updated: 2021/10/06 16:40:12 by hohnuki          ###   ########.fr       */
+/*   Updated: 2021/10/07 18:32:31 by hohnuki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memmove(void *s1, const void *s2, int size)
+#include "libft.h"
+
+void	*ft_memmove(void *s1, const void *s2, size_t size)
 {
 	int				i;
 	unsigned char	*ptr1;
@@ -19,7 +21,7 @@ void	*ft_memmove(void *s1, const void *s2, int size)
 	i = 0;
 	ptr1 = (unsigned char *)s1;
 	ptr2 = (unsigned char *)s2;
-	while (i < size)
+	while (i < (int)size)
 	{
 		ptr1[i] = ptr2[i];
 		i++;
