+++ ft_memset.c
@@ -6,11 +6,13 @@
 /*   By: hohnuki <hohnuki@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/06 14:47:57 by hohnuki           #+#    #+#             */
-/*   Updated: 2021/10/06 15:31:28 by hohnuki          ###   ########.fr       */
+/*   Updated: 2021/10/07 18:23:49 by hohnuki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *str, int c, int size)
+#include "libft.h"
+
+void	*ft_memset(void *str, int c, size_t size)
 {
 	int				i;
 	unsigned char	*ptr;
@@ -19,7 +21,7 @@ void	*ft_memset(void *str, int c, int size)
 	i = 0;
 	ptr = (unsigned char *)str;
 	value = (unsigned char)c;
-	while (i < size)
+	while (i < (int)size)
 	{
 		ptr[i] = value;
 		i++;
