+++ ft_strlcpy.c
@@ -6,27 +6,29 @@
 /*   By: hohnuki <hohnuki@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/06 16:41:23 by hohnuki           #+#    #+#             */
-/*   Updated: 2021/10/06 17:04:21 by hohnuki          ###   ########.fr       */
+/*   Updated: 2021/10/07 18:51:12 by hohnuki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlcpy(char *dest, char *src, unsigned int size)
+#include "libft.h"
+
+size_t	ft_strlcpy(char *dest, char *src, size_t size)
 {
-	unsigned int	i;
-	int				count;
+	int	i;
+	int	count;
 
 	i = 0;
 	count = 0;
-	while (i < (size - 1) && src[i] != '\0')
+	while (i < (int)(size - 1) && src[i] != '\0')
 	{
 		dest[i] = src[i];
 		i++;
 		count++;
 	}
-	while (i < size)
+	while (i < (int)size)
 	{
 		dest[i] = '\0';
 		i++;
 	}
-	return (count);
+	return ((size_t)count);
 }
